+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Changes to the extension.
 - When running a game in VS Code, error modals are now shown in VS Code instead of being silently swallowed.
 - SugarCube: Improved syntax highlighting of twee links.
 - SugarCube: Fixed parsing of special variables in `<<script>>` macros.
+- SugarCube: No longer incorrectly indexes apparent variables in non-data HTML attributes.
 
 ## [1.1.5] - 2026 04 12